Fix for off-screen windows when restoring
Some users reported that Ameol restores windows off-screen as a result
of huge unsigned integer values being written to the registry. This
seems caused by either GetWindowPlacement returning an error or the top
left corner being off the edge in which case the negative coordinates
are saved unsigned and treated as unsigned when read back. So now check
the return value from GetWindowPlacement and save them as signed values.